fix: extract useWalletConnect hook and silence cancellation errors
